backend/requirements.txt
Django>=5.0
django-ninja>=1.1.0
gunicorn>=21.0.0
PyYAML
django-ninja-extra